Fix phpunit errors, warnings and deprecations
