+++ tests/Integration/MessageHandler/Connection/ConnectionCreateHandlerTest.php
@@ -24,6 +24,7 @@ final class ConnectionCreateHandlerTest extends KernelTestCase
         parent::setUp();
         $this->channelRepository = self::getContainer()->get('sylius.repository.channel');
 
+        /** @var \Fidry\AliceDataFixtures\Loader\PurgerLoader $fixtureLoader */
         $fixtureLoader = self::getContainer()->get('fidry_alice_data_fixtures.loader.doctrine');
         $fixtureLoader->load([
             self::FIXTURE_BASE_DIR . '/channels.yaml',
