+++ tests/Integration/MessageHandler/Contact/ContactCreateHandlerTest.php
@@ -25,6 +25,7 @@ final class ContactCreateHandlerTest extends KernelTestCase
         parent::setUp();
         $this->customerRepository = self::getContainer()->get('sylius.repository.customer');
 
+        /** @var \Fidry\AliceDataFixtures\Loader\PurgerLoader $fixtureLoader */
         $fixtureLoader = self::getContainer()->get('fidry_alice_data_fixtures.loader.doctrine');
         $fixtureLoader->load([
             self::FIXTURE_BASE_DIR . '/customers.yaml',
