+++ tests/Integration/MessageHandler/EcommerceCustomer/EcommerceCustomerCreateHandlerTest.php
@@ -34,6 +34,7 @@ final class EcommerceCustomerCreateHandlerTest extends KernelTestCase
         $this->customerRepository = self::getContainer()->get('sylius.repository.customer');
         $this->channelRepository = self::getContainer()->get('sylius.repository.channel');
 
+        /** @var \Fidry\AliceDataFixtures\Loader\PurgerLoader $fixtureLoader */
         $fixtureLoader = self::getContainer()->get('fidry_alice_data_fixtures.loader.doctrine');
         $fixtureLoader->load([
             self::FIXTURE_BASE_DIR . '/channels.yaml',
